Organized nvi-hooks and removed redundant type checking on year as number
