+++ src/components/nvi/hooks/tests/useNviInstitutionStatusNumbers.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from 'vitest';
-import { ApprovalStatusAggregation } from '../../../types/nvi.types';
-import { getTotalResults } from './useNviInstitutionStatusNumbers';
+import { ApprovalStatusAggregation } from '../../../../types/nvi.types';
+import { getTotalResults } from '../useNviInstitutionStatusNumbers';
 
 const makeApprovalStatus = (overrides: Partial<ApprovalStatusAggregation> = {}): ApprovalStatusAggregation => ({
   New: 0,
